make graphql exception handler more simple and readable
