--- src/main/java/com/irdaislakhuafa/garbagepickupapi/exceptions/gql/GraphQLDataAlreadyExistsHandler.java
@@ -1,28 +1,0 @@
-package com.irdaislakhuafa.garbagepickupapi.exceptions.gql;
-
-import org.springframework.graphql.execution.DataFetcherExceptionResolverAdapter;
-import org.springframework.graphql.execution.ErrorType;
-import org.springframework.stereotype.Component;
-
-import com.irdaislakhuafa.garbagepickupapi.exceptions.custom.DataAlreadyExists;
-
-import graphql.GraphQLError;
-import graphql.GraphqlErrorBuilder;
-import graphql.schema.DataFetchingEnvironment;
-
-@Component
-public class GraphQLDataAlreadyExistsHandler extends DataFetcherExceptionResolverAdapter {
-    @Override
-    protected GraphQLError resolveToSingleError(Throwable ex, DataFetchingEnvironment env) {
-        if (ex instanceof DataAlreadyExists) {
-            return GraphqlErrorBuilder.newError()
-                    .errorType(ErrorType.BAD_REQUEST)
-                    .message(ex.getMessage())
-                    .path(env.getExecutionStepInfo().getPath())
-                    .location(env.getField().getSourceLocation())
-                    .build();
-        }
-
-        return null;
-    }
-}
